refactor: Move `isSpinnerPopup` function to `SpinnerAdapterAssertions`
This commit moves the `isSpinnerPopup` function into the `SpinnerAdapterAssertions` to avoid use of static functions
